ref : project architecture

+++ src/main/java/com/example/planeo_back/domain/models/ExpenseByTag.java
@@ -1,6 +1,6 @@
 package com.example.planeo_back.domain.models;
 
-import com.example.planeo_back.domain.entity.enums.Tag;
+import com.example.planeo_back.domain.enums.Tag;
 
-public record ExpenseByTag(Tag tag, long total) {
+public record ExpenseByTag(Tag tag, Double total) {
 }
