+++ src/main/java/com/example/planeo_back/web/DTO/expense/ExpensePerMonthDTO.java
@@ -1,4 +1,4 @@
 package com.example.planeo_back.web.DTO.expense;
 
-public record ExpensePerMonthDTO(int month, int amount) {
+public record ExpensePerMonthDTO(Integer month, Double amount) {
 }
